Fix ambiguous call to abs
Fixes compilation error:

```
vlasovsolver/cpu_trans_map_amr.cpp:613:89: virhe: call of overloaded ”abs(std::array<long unsigned int, 3>::value_type)” is ambiguous
```

+++ vlasovsolver/cpu_trans_map_amr.cpp
@@ -610,7 +610,7 @@ void getSeedIds(const dccrg::Dccrg<SpatialCell,dccrg::Cartesian_Geometry>& mpiGr
 
             // If a neighbor is non-local or across a periodic boundary, then we use this
             // cell as a seed for pencils
-            if ( abs ( myIndices[dimension] - nbrIndices[dimension] ) >
+            if ( abs ( (int64_t)(myIndices[dimension] - nbrIndices[dimension]) ) >
                  pow(2,mpiGrid.get_maximum_refinement_level()) ||
                  !mpiGrid.is_local(nbrPair.first)) {
